apps/worker/requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
apify-client
replicate
prisma
python-dotenv
cryptography
apscheduler
